requirements.txt
discord
colorama
tortoise-orm
PyPika
